requirements.txt
pandas>=2.0.0
numpy>=1.21.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
networkx>=3.0
torch>=2.0.0
torch-geometric>=2.3.0
scikit-learn>=1.3.0
openai>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-dotenv>=1.0.0
pydantic>=2.0.0
streamlit>=1.28.0
plotly>=5.15.0
python-multipart>=0.0.6
